+++ src/components/spartan/SAvatar/types.ts
@@ -1,0 +1,9 @@
+export type TAvatarProps = {
+    class: string;
+    borderless: boolean;
+    indicator: boolean;
+    indicatorPosition: 'left-top' | 'left-bottom' | 'right-top' | 'right-bottom';
+    name: string;
+    size: 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl';
+    src: string;
+};